# Document/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AxioDB Dashboard HTTP API (Control Server)",
    "description": "REST API behind the AxioDB Dashboard - the web GUI AxioDB starts on port 27018 - for managing databases, collections, indexes, documents, users, and roles.\n\nThis is not a hosted service. It runs inside your own AxioDB process: enable it with `new AxioDB({ GUI: true })`, or with `AXIODB_GUI=true` on the `theankansaha/axiodb` Docker image. https://axiodb.in serves documentation only.\n\nThis API is one of several AxioDB surfaces and is not interchangeable with the others:\n- Embedded library (`npm install axiodb`) - in-process, no HTTP at all. https://axiodb.in/api-reference\n- AxioDBCloud TCP server/client on port 27019 - its own wire protocol, not HTTP. https://axiodb.in/cloud\n- MCP server on port 27020 (Docker image only) - Streamable HTTP at /mcp, 32 tools. https://axiodb.in/mcp-server\n\nThe Dashboard, AxioDBCloud authentication, and the MCP server share one user and role store.\n\nHuman-readable reference: https://axiodb.in/server-api",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "contact": {
      "name": "AxioDB",
      "url": "https://axiodb.in"
    }
  },
  "servers": [
    {
      "url": "http://localhost:27018",
      "description": "AxioDB Dashboard on the machine running AxioDB (port is fixed; remap at the Docker layer if needed)"
    }
  ],
...
  "paths": {
...
    "/api/system": {
      "get": {
        "tags": [
          "System & Health"
        ],
        "summary": "Operational detail for the Status page: process, memory, cache, instance and service ports",
        "description": "Operational detail for the Status page: process, memory, cache, instance and service ports. Requires authentication and the dashboard:view permission, because it reports host characteristics that should not be exposed anonymously.",
        "operationId": "getApiSystem",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success - Returns system information",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 200,
                  "status": "success",
                  "message": "System information",
                  "data": {
                    "process": {
                      "uptimeSeconds": 86400,
                      "startedAt": "2025-10-30T12:00:00.000Z",
                      "pid": 4821,
                      "nodeVersion": "v20.11.0",
                      "platform": "linux",
                      "arch": "x64",
                      "cpuCount": 8
                    },
                    "memory": {
                      "rss": 92.4,
                      "heapUsed": 41.2,
                      "heapTotal": 63.5,
                      "external": 2.1,
                      "unit": "MB"
                    },
                    "cache": {
                      "used": 3.42,
                      "max": 512,
                      "unit": "MB"
                    },
                    "instance": {
                      "version": "15.0.0",
                      "rootName": "AxioDB",
                      "path": "/app/AxioDB",
                      "totalDatabases": 3
                    },
                    "services": [
                      {
                        "name": "Dashboard + HTTP API",
                        "port": 27018,
                        "running": true
                      },
                      {
                        "name": "AxioDBCloud TCP",
                        "port": 27019,
                        "running": false
                      },
                      {
                        "name": "MCP server",
                        "port": 27020,
                        "running": false,
                        "note": "Docker image only"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Session invalid or expired"
          },
          "403": {
            "description": "Forbidden - Role lacks dashboard:view"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/dashboard-stats": {
      "get": {
        "tags": [
          "System & Health"
        ],
        "summary": "Retrieves dashboard statistics including total databases, collections, documents, and storage usage",
        "description": "Retrieves dashboard statistics including total databases, collections, documents, and storage usage. Perfect for admin dashboards.",
        "operationId": "getApiDashboardStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success - Returns dashboard statistics",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 200,
                  "status": "success",
                  "message": "Dashboard statistics",
                  "data": {
                    "totalDatabases": 5,
                    "totalCollections": 12,
                    "totalDocuments": 1245,
                    "totalSize": "15.4 MB"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/auth/users": {
      "get": {
        "tags": [
          "Authentication & Access Control"
        ],
        "summary": "Lists all users (Super Admin only - requires the `user:view` permission)",
        "description": "Lists all users (Super Admin only - requires the `user:view` permission). Password hashes are never included in the response.",
        "operationId": "getApiAuthUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success - Returns the user list",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 200,
                  "message": "List of Users",
                  "data": [
                    {
                      "username": "admin",
                      "role": "Super Admin",
                      "mustChangePassword": false,
                      "isActive": true,
                      "createdAt": "2026-01-01T00:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Session invalid or expired"
          },
          "403": {
            "description": "Forbidden - Missing user:view permission"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "Authentication & Access Control"
        ],
        "summary": "Creates a new user with a temporary password and an assigned role (Super Admin only - requires `user:create`)",
        "description": "Creates a new user with a temporary password and an assigned role (Super Admin only - requires `user:create`). The new user is forced to change their password on first login.",
        "operationId": "postApiAuthUsers",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "New account's username"
                  },
                  "password": {
                    "type": "string",
                    "description": "Temporary password"
                  },
                  "role": {
                    "type": "string",
                    "description": "One of the existing role names"
                  }
                },
                "required": [
                  "username",
                  "password",
                  "role"
                ]
              },
              "example": {
                "username": "jane",
                "password": "TempPass123",
                "role": "Admin"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created - User created successfully",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 201,
                  "message": "User created successfully",
                  "data": {
                    "username": "jane",
                    "role": "Admin"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid username/password or unknown role"
          },
          "401": {
            "description": "Unauthorized - Session invalid or expired"
          },
          "403": {
            "description": "Forbidden - Missing user:create permission"
          },
          "409": {
            "description": "Conflict - Username already exists"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/auth/roles": {
      "get": {
        "tags": [
          "Authentication & Access Control"
        ],
        "summary": "Lists all roles, including custom roles created from the permission catalogue (Super Admin only - requires `role:view`).",
        "description": "Lists all roles, including custom roles created from the permission catalogue (Super Admin only - requires `role:view`).",
        "operationId": "getApiAuthRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success - Returns the role list",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 200,
                  "message": "List of Roles",
                  "data": [
                    {
                      "roleName": "Super Admin",
                      "permissions": [
                        "..."
                      ],
                      "isSystemRole": true,
                      "createdAt": "2026-01-01T00:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Session invalid or expired"
          },
          "403": {
            "description": "Forbidden - Missing role:view permission"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "Authentication & Access Control"
        ],
        "summary": "Creates a new role from the predefined permission catalogue (Super Admin only - requires `role:create`)",
        "description": "Creates a new role from the predefined permission catalogue (Super Admin only - requires `role:create`). Every permission key must exist in the catalogue returned by GET /api/auth/roles/permissions.",
        "operationId": "postApiAuthRoles",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roleName": {
                    "type": "string",
                    "description": "Unique name for the new role"
                  },
                  "permissions": {
                    "type": "array",
                    "items": {},
                    "description": "Array of permission keys from the catalogue"
                  }
                },
                "required": [
                  "roleName",
                  "permissions"
                ]
              },
              "example": {
                "roleName": "Auditor",
                "permissions": [
                  "document:view",
                  "document:query",
                  "dashboard:view"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created - Role created successfully",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 201,
                  "message": "Role created successfully",
                  "data": {
                    "roleName": "Auditor",
                    "permissions": [
                      "document:view",
                      "document:query",
                      "dashboard:view"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Unknown permission key(s)"
          },
          "401": {
            "description": "Unauthorized - Session invalid or expired"
          },
          "403": {
            "description": "Forbidden - Missing role:create permission"
          },
          "409": {
            "description": "Conflict - Role name already exists"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "axiodb_session",
        "description": "httpOnly session cookie issued by POST /api/auth/login. RBAC-enforced per endpoint."
      }
    }
  }
}